# Conflicts:
#	src/main/java/socialnet/dto/CommonRs.java
#	src/main/java/socialnet/dto/CurrencyRs.java
#	src/main/java/socialnet/repository/PostRepository.java
#	src/main/java/socialnet/repository/TagRepository.java
#	src/main/java/socialnet/service/PostService.java

+++ src/main/java/socialnet/model/Post.java
@@ -6,7 +6,7 @@ import java.sql.Timestamp;
 
 @Data
 public class Post {
-    private Integer id;
+    private Long id;
 
     private Boolean isBlocked;
 
@@ -20,5 +20,5 @@ public class Post {
 
     private String title;
 
-    private Integer authorId;
+    private Long authorId;
 }
